fix: Force bash for non-standard flags to exec
This addresses a regression introduced by #224. POSIX sh does not
define arguments to the exec built-in, resulting in an error when
using 'exec -a' to set the process name (argv[0]).

This change executes /bin/bash instead of sh, and also changes
to use 'exec -l', which automatically handles the hyphen prefix
required to trigger login shell behavior.

This also explicitly runs /bin/sh, aligning the behavior of
"coder sh" and the frontend Terminal application.